Vanha React frontti poistettu, uuden vite projektin kansiorakennetta muutettu
