+++ frontend/src/App.jsx
@@ -6,17 +6,17 @@ import { ThemeProvider, useTheme } from './ThemeProvider';
 import '@css/styles.css';
 import Header from '@components/header/Header';
 import Footer from '@components/footer/Footer';
-import Home from '@pages/Home';
+import Home from '@content/homepage/Home';
 // importtaa muut sivut
 
 function App() {
   const { theme, toggleTheme } = useTheme();
-  const [ user, setUser ] = useState(null)
+  const [user, setUser] = useState(null)
 
   useEffect(() => {
     const savedTheme = localStorage.getItem('theme');
     if (savedTheme) {
-      toggleTheme(savedTheme); 
+      toggleTheme(savedTheme);
     }
   }, []);
 
@@ -31,14 +31,14 @@ function App() {
 
   return (
     <>
-    <Router>
-      <ThemeProvider>
-        <div className={`body ${theme}`}>
-          <Header user={user} handleLogout={handleLogout} />
-          
-          <Routes>
-            <Route path="/" exact element={<Home />} />
-          {/****** Loput routet, esim.
+      <Router>
+        <ThemeProvider>
+          <div className={`body ${theme}`}>
+            <Header user={user} handleLogout={handleLogout} />
+
+            <Routes>
+              <Route path="/" exact element={<Home />} />
+              {/****** Loput routet, esim.
           <Route path="/search" element={<Search />} />
           <Route path="/community" element={<Community />} />
           <Route path="/login" element={<Login setUser={setUser}/>} />
@@ -47,13 +47,13 @@ function App() {
           <Route path="/group/" element={<GroupDetails/>} />
           <Route path="/profile/" element={<ProfileDetails/>} />
           ********/}
-          </Routes>
+            </Routes>
 
-          <Footer toggleTheme={toggleTheme} theme={theme} />
-        </div>
+            <Footer toggleTheme={toggleTheme} theme={theme} />
+          </div>
 
-      </ThemeProvider>
-    </Router>
+        </ThemeProvider>
+      </Router>
     </>
   )
 }
