--- frontend/src/pages/Home.jsx
@@ -1,77 +1,0 @@
-import React from 'react';
-import Comingsoon from '@content/Comingsoon';
-
-const Home = () => {
-
-  return (
-
-    <div className="content">
-
-    <h2>Tulossa elokuvateattereihin </h2>
-
-    {<Comingsoon />}
-      
-    <h2>H2 Some Title Here </h2>
-
-    <p>Put some content here</p>
-
-    <p>Huom. Tämä malli ei ole responsiivinen. Ts. "hajoaa" pienemmillä näytöillä. </p>
-
-    <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. 
-    Urna porttitor rhoncus dolor purus non enim praesent. Turpis nunc eget lorem dolor sed. Ut tristique et egestas quis ipsum suspendisse ultrices gravida. 
-    Et ultrices neque ornare aenean euismod elementum nisi quis. Tristique et egestas quis ipsum suspendisse ultrices gravida dictum. Habitasse platea dictumst quisque sagittis purus sit. 
-    Amet massa vitae tortor condimentum lacinia quis. Amet consectetur adipiscing elit pellentesque habitant. 
-    Aliquet bibendum enim facilisis gravida neque convallis a cras semper. In nulla posuere sollicitudin aliquam ultrices sagittis orci a. 
-    Nisl purus in mollis nunc. Sit amet facilisis magna etiam tempor orci. Sed tempus urna et pharetra pharetra. Diam quam nulla porttitor massa id. 
-    Pulvinar sapien et ligula ullamcorper malesuada proin libero nunc consequat. Cras fermentum odio eu feugiat pretium nibh ipsum consequat nisl.</p>
-
-
-    <p>Ultrices mi tempus imperdiet nulla malesuada pellentesque. Iaculis nunc sed augue lacus viverra vitae congue eu consequat. 
-    Gravida arcu ac tortor dignissim convallis aenean et. Hac habitasse platea dictumst vestibulum rhoncus. Suspendisse interdum consectetur libero id faucibus. 
-    Iaculis eu non diam phasellus vestibulum. Sed faucibus turpis in eu mi. A condimentum vitae sapien pellentesque habitant morbi tristique senectus et. Etiam sit amet nisl purus. 
-    Sed cras ornare arcu dui vivamus arcu felis bibendum. Vitae suscipit tellus mauris a diam maecenas. Viverra tellus in hac habitasse. Elementum eu facilisis sed odio morbi quis commodo odio. 
-    Purus non enim praesent elementum facilisis. Id aliquet risus feugiat in. Enim nec dui nunc mattis enim ut.</p>
-
-    <p>Nibh sit amet commodo nulla facilisi nullam. Nibh tellus molestie nunc non blandit massa enim. Libero nunc consequat interdum varius sit amet mattis vulputate enim. 
-    Fames ac turpis egestas maecenas pharetra convallis posuere. Ornare lectus sit amet est placerat in egestas erat imperdiet. Lectus nulla at volutpat diam ut venenatis tellus. 
-    Purus non enim praesent elementum. Urna porttitor rhoncus dolor purus non enim praesent elementum. Id eu nisl nunc mi ipsum faucibus vitae aliquet. Eu nisl nunc mi ipsum faucibus vitae aliquet nec. 
-    Neque volutpat ac tincidunt vitae semper quis lectus. Leo vel orci porta non pulvinar neque laoreet. Porttitor rhoncus dolor purus non enim praesent. Lectus nulla at volutpat diam ut venenatis.</p>
-
-    <p>Purus sit amet volutpat consequat mauris nunc congue. Amet volutpat consequat mauris nunc congue. Aliquam faucibus purus in massa tempor. Ultrices gravida dictum fusce ut placerat. 
-    Amet commodo nulla facilisi nullam vehicula ipsum a. Purus faucibus ornare suspendisse sed nisi lacus sed viverra tellus. Semper auctor neque vitae tempus. 
-    Mattis nunc sed blandit libero volutpat sed cras ornare. Enim facilisis gravida neque convallis. Facilisis sed odio morbi quis. Quis commodo odio aenean sed adipiscing diam donec adipiscing tristique. 
-    Est ullamcorper eget nulla facilisi. Sollicitudin ac orci phasellus egestas tellus rutrum tellus pellentesque eu. Id volutpat lacus laoreet non. Mi sit amet mauris commodo quis. </p>
-    Tortor dignissim convallis aenean et.
-
-    <h2>H2 Some Title Here </h2>
-
-    <p>Morbi non arcu risus quis varius quam. Urna duis convallis convallis tellus. Adipiscing diam donec adipiscing tristique risus nec. Lacinia at quis risus sed vulputate odio ut enim. 
-    Etiam sit amet nisl purus in mollis nunc sed id. Sapien et ligula ullamcorper malesuada proin. Ut morbi tincidunt augue interdum velit euismod in pellentesque. Sed nisi lacus sed viverra tellus. 
-    Aenean sed adipiscing diam donec adipiscing tristique risus. Quis varius quam quisque id. Non quam lacus suspendisse faucibus interdum. Nunc sed augue lacus viverra vitae. 
-    Quam pellentesque nec nam aliquam sem et tortor. Molestie nunc non blandit massa enim nec dui nunc mattis. Pellentesque sit amet porttitor eget dolor morbi non arcu risus. 
-    Volutpat odio facilisis mauris sit amet massa.</p>
-
-      
-    <p>Nibh sit amet commodo nulla facilisi nullam. Nibh tellus molestie nunc non blandit massa enim. Libero nunc consequat interdum varius sit amet mattis vulputate enim. 
-    Fames ac turpis egestas maecenas pharetra convallis posuere. Ornare lectus sit amet est placerat in egestas erat imperdiet. Lectus nulla at volutpat diam ut venenatis tellus. 
-    Purus non enim praesent elementum. Urna porttitor rhoncus dolor purus non enim praesent elementum. Id eu nisl nunc mi ipsum faucibus vitae aliquet. Eu nisl nunc mi ipsum faucibus vitae aliquet nec. 
-    Neque volutpat ac tincidunt vitae semper quis lectus. Leo vel orci porta non pulvinar neque laoreet. Porttitor rhoncus dolor purus non enim praesent. Lectus nulla at volutpat diam ut venenatis.</p>
-
-    <p>Purus sit amet volutpat consequat mauris nunc congue. Amet volutpat consequat mauris nunc congue. Aliquam faucibus purus in massa tempor. Ultrices gravida dictum fusce ut placerat. 
-    Amet commodo nulla facilisi nullam vehicula ipsum a. Purus faucibus ornare suspendisse sed nisi lacus sed viverra tellus. Semper auctor neque vitae tempus. 
-    Mattis nunc sed blandit libero volutpat sed cras ornare. Enim facilisis gravida neque convallis. Facilisis sed odio morbi quis. Quis commodo odio aenean sed adipiscing diam donec adipiscing tristique. 
-    Est ullamcorper eget nulla facilisi. Sollicitudin ac orci phasellus egestas tellus rutrum tellus pellentesque eu. Id volutpat lacus laoreet non. Mi sit amet mauris commodo quis. </p>
-    Tortor dignissim convallis aenean et.
-
-    <p>Morbi non arcu risus quis varius quam. Urna duis convallis convallis tellus. Adipiscing diam donec adipiscing tristique risus nec. Lacinia at quis risus sed vulputate odio ut enim. 
-    Etiam sit amet nisl purus in mollis nunc sed id. Sapien et ligula ullamcorper malesuada proin. Ut morbi tincidunt augue interdum velit euismod in pellentesque. Sed nisi lacus sed viverra tellus. 
-    Aenean sed adipiscing diam donec adipiscing tristique risus. Quis varius quam quisque id. Non quam lacus suspendisse faucibus interdum. Nunc sed augue lacus viverra vitae. 
-    Quam pellentesque nec nam aliquam sem et tortor. Molestie nunc non blandit massa enim nec dui nunc mattis. Pellentesque sit amet porttitor eget dolor morbi non arcu risus. 
-    Volutpat odio facilisis mauris sit amet massa.</p>
-
-    </div>
-
-  );
-};
-
-export default Home;
