--- frontend/src/src-original/css/comingsoon.css
@@ -1,70 +1,0 @@
-.event-list {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center; /* Keskitä sisällön vaakasuunnassa */
-  }
-
-  .events-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin: 10vmin;
-    font-family: sans-serif;
-    overflow: hidden;
-    transform: skew(5deg);
-  }
-
-  .event-item {
-    flex: 1;
-    transition: all 1s ease-in-out;
-    height: 75vmin;
-    position: relative;
-  }
-  /*
-  .event-item:not(:nth-child(5)){
-    margin-right: 1em;
-  }
-*/
-  .event-item img {
-    width: 100%; 
-    height: 100%; 
-    object-fit: cover;
-    transition: all 1s ease-in-out;
-    filter: grayscale(100%)
-  }
-
-
-  .head {
-    color:black;
-    background-color: white;
-    padding: 0.5em;
-    transform: rotate(-90deg);
-    transform-origin: 0% 0%;
-    transition: all 0.5s ease-in-out;
-    min-width: 100%;
-    text-align: center;
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    font-size: 1em;
-    white-space: nowrap;
-  }
-
-
-  .event-item:hover {
-    flex-grow: 10;
-  }
-
-
-  .event-item:hover img {
-    filter: grayscale(0);
-  }
-
-
-  .event-item:hover .head {
-    text-align: center;
-    top: calc(100%-2em);
-    color: white;
-    background: black;
-    transform: rotate(0deg);
-  }
