+++ frontend/vite.config.js
@@ -7,7 +7,6 @@ export default defineConfig({
   plugins: [react()],
   resolve: {
     alias: {
-      '@pages': path.resolve(__dirname, 'src/pages'),
       '@components': path.resolve(__dirname, 'src/components'),
       '@content': path.resolve(__dirname, 'src/components/content'),
       '@css': path.resolve(__dirname, 'src/css'),
