--- leffasovellus/src/App.js
@@ -1,13 +1,0 @@
-import './App.css';
-import Comingsoon from './components/content/Comingsoon';
-
-function App() {
-  return (
-    <div className="App">
-      <h1>Leffasovellus</h1>
-      <Comingsoon />
-    </div>
-  );
-}
-
-export default App;
