--- leffasovellus/src/css/styles.css
@@ -1,174 +1,0 @@
-:root {
-    --dark-background-color: #333;
-    --light-background-color: #f5f5f5;
-    --dark-foreground-color: #000;
-    --light-foreground-color: #fff;
-    --grey-foreground-color: #ccc;
-    --accent-color: #6D0509;
-    --general-padding: 2em;
-    --general-margin: 1em;
-}
-
-* {
-    margin: 0;
-    padding: 0;
-    color: var(--dark-foreground-color);
-    font-family: Arial, Helvetica, sans-serif;
-    overflow-x: hidden; /* Estää horisontaalisen rullauksen */
-    overflow-y: auto; /* Sallii vertikaalisen rullauksen */
-}
-
-*,*::before,*::after {
-    box-sizing: border-box;
-}
-
-a {
-    text-decoration: none;
-}
-
-a:hover {
-    color: var(--accent-color);
-    font-weight: bold;
-}
-
-div#container {
-    display: flex;
-    flex-direction: column;
-    margin: var(--general-margin);
-}
-
-div#content {
-    display: inherit;
-}
-
-div#content section {
-    width: 80vw;
-    display: inherit;
-    flex-direction: row;
-    flex-wrap: wrap;
-}
-
-div#content aside {
-    width: 20vw;
-}
-
-article {
-    flex: 33.33%;
-}
-
-article div {
-    margin: var(--general-margin);
-    border: 1px solid var(--grey-foreground-color);
-}
-
-article div img {
-    height: 500px;
-    width: 100%;
-    object-fit: cover;
-    object-position: center;
-}
-
-header {
-    display: inherit;
-    justify-content: center;
-    align-items: center;
-    height: 150px;
-}
-
-header h1 {
-    font-family: Verdana, Geneva, Tahoma, sans-serif;
-}
-
-nav {
-    display: inherit;
-    align-items: center;
-    border-top: 1px solid var(--dark-foreground-color);
-    border-bottom: 1px solid var(--dark-foreground-color);
-    height: 50px;
-    width: auto;
-}
-
-nav ul {
-    padding-left: calc(var(--general-padding)/2);
-    background-color: #161515;
-}
-
-nav li {
-    display: inline-block;
-    margin-right: var(--general-margin);
-    color: #FFF;
-    background-color: #161515;
-}
-
-nav li a {
-    font-family: Arial, sans-serif, Helvetica;
-    display: inline-block;
-    color: #FFF;
-    text-decoration: none;
-    padding: 0.5em 1em;
-}
-
-nav li a:hover {
-    background-color: #111;
-    color: #FFF;
-    font-weight: bold;
-}
-
-aside {
-    padding: var(--general-padding);
-    padding-top: calc(var(--general-padding)*1.5);
-}
-
-footer {
-    background-color: var(--dark-background-color);
-    font-family: Arial, sans-serif, Helvetica;
-    display: inherit;
-    justify-content: center;
-    align-items: center;
-    height: 50px;
-    width: auto;
-    position: fixed;
-    bottom: 0;
-    width: 100%;
-    text-align: center;
-    padding: 10px 0; 
-}
-
-footer p {
-    color: var(--light-foreground-color);
-    font-family: Arial, sans-serif, Helvetica;
-    font-weight: bold;
-    justify-content: center;
-}
-
-@media only screen and (max-width: 1200px) {
-    div#content {
-        flex-direction: column;
-    }
-
-    div#content section {
-        width: 100vw;
-        flex-direction: row;
-    }
-
-    div#content aside {
-        width: 100vw;
-        padding-top: var(--general-padding);
-        padding-left: calc(var(--general-padding)/2);
-    }
-
-    article {
-        flex: 50%;
-    }
-
-    article div img {
-        height: 350px;
-    }
-}
-
-@media only screen and (max-width: 992px) {
-    div#content section {
-        width: 100vw;
-        flex-direction: column;
-    }
-}
