--- leffasovellus/src/pages/main.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import App from './App.js';
-import './css/styles.css';
-
-ReactDOM.createRoot(document.getElementById('root')).render(
-  <React.StrictMode>
-      <App />
-  </React.StrictMode>
-);
